better post-patch clean-up; remains to handle memory, transition-conditionals.
Modified small2.world to make it more reasonable, because the previous
edition (which broke my reachable-set method) was degenerate in the
sense that it reduced to the global problem, i.e. only realized for
abstract neighborhood covering the entire space.  As of this commit,
you can try both example_data/small.world and small2.world,

./dgridworld_example.py example_data/small2{,_real}.world